Fixed: Some public modules weren't recognized
Dojo docs has a hard time with documentation on declared
objects, so this leads to some of our public
modules to not get properly tagged as such. There are only six of them,
so they are added to publicModulesWorkarounds.json for now.

Also refactored out the namespaces to it's own JSON file:
namespaces.json